requirements.txt
absl-py==2.0.0
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
attrs==23.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==5.3.2
# certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
certifi
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cycler==0.11.0
decorator==5.1.1
dlib==19.24.2
exceptiongroup==1.2.0
facenet-pytorch==2.5.3
fastapi==0.103.2
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
hsemotion==0.3.0
hsemotion-onnx==0.3.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.3
jsonschema==4.17.3
kiwisolver==1.4.5
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
mdurl==0.1.2
# mediapipe==0.8.9
mpmath==1.3.0
numpy==1.21.6
onnx==1.14.1
onnxruntime==1.14.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==12.0.1
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
pyparsing==3.1.1
pyreadline==2.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyUserInput==0.1.10
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
safetensors==0.4.1
scipy==1.7.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
streamlit==1.23.1
sympy==1.10.1
tenacity==8.2.3
timm==0.9.12
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.4
tzlocal==4.3.1
urllib3==2.0.7
uvicorn==0.22.0
validators==0.20.0
watchdog==3.0.0
wincertstore==0.2
zipp==3.15.0
